refactor: move emit methods to build
